requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1.1
django-cleanup==5.1.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-markdownx==3.0.1
filelock==3.0.12
Markdown==3.2.2
pbr==5.5.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.2
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
